Expenses

+++ db/migrate/20210717121336_create_households.rb
@@ -2,7 +2,7 @@ class CreateHouseholds < ActiveRecord::Migration[6.0]
   def change
     create_table :households do |t|
       t.string :name
-      t.numeric :household_member_id
+      t.integer :household_member_id
       t.timestamps
     end
   end
